move redeem into components
